@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <GenerateProgramFile>false</GenerateProgramFile>
     </PropertyGroup>
 
@@ -9,10 +9,10 @@
         <ProjectReference Include="..\..\..\..\src\Miru.PageTesting.Firefox\Miru.PageTesting.Firefox.csproj" />
         <ProjectReference Include="..\..\..\..\src\Miru.PageTesting\Miru.PageTesting.csproj" />
 
-        <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
+        <PackageReference Include="NUnit3TestAdapter" Version="4.3.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
-        <PackageReference Include="Selenium.WebDriver" Version="4.5.0" />
-        <PackageReference Include="Selenium.WebDriver.GeckoDriver" Version="0.31.0.1" />
+        <PackageReference Include="Selenium.WebDriver" Version="4.6.0" />
+        <PackageReference Include="Selenium.WebDriver.GeckoDriver" Version="0.32.0" />
     </ItemGroup>
     
     <ItemGroup>
