@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <IsPackable>true</IsPackable>
         <PackAsTool>true</PackAsTool>
         <PackageId>Miru.Cli</PackageId>
