@@ -16,7 +16,7 @@ public class MiniProfilerBehavior<TRequest, TResponse> : IPipelineBehavior<TRequ
         _options = options.Value;
     }
 
-    public async Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
     {
         using (MiniProfiler.Current.Step($"Handling {request.GetType()}"))
         {
