@@ -17,8 +17,8 @@ public class AuthorizationBehavior<TRequest, TResponse> :
 
     public async Task<TResponse> Handle(
         TRequest request, 
-        CancellationToken cancellationToken, 
-        RequestHandlerDelegate<TResponse> next)
+        RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
     {
         var featureInfo = new FeatureInfo(typeof(TRequest));
             
